Add help menu

+++ omzsh-install.sh
@@ -1,5 +1,18 @@
 #!/bin/sh
 
+if [ $1 ]; then
+	echo " Install Script for zshfish theme"
+	echo ""
+	echo " How to use:"
+	echo "   Run this script and zshfish thene will be automativally installed"
+	echo "   Requires omzsh"
+	echo ""
+	echo " Options:"
+	echo "   --help    Display this screen and exit"
+	echo ""
+	exit 0
+fi
+
 rm -f ~/.oh-my-zsh/custom/themes/zshfish.zsh-theme
 cp zshfish.zsh-theme ~/.oh-my-zsh/custom/themes
